moved stuff around so a bit less crap
