+++ test/Ocelot.ManualTest/Startup.cs
@@ -1,15 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using CacheManager.Core;
-using IdentityServer4.AccessTokenValidation;
-using IdentityServer4.Models;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Ocelot.Configuration.Provider;
 using Ocelot.DependencyInjection;
 using Ocelot.Middleware;
 using ConfigurationBuilder = Microsoft.Extensions.Configuration.ConfigurationBuilder;
@@ -18,8 +13,6 @@ namespace Ocelot.ManualTest
 {
     public class Startup
     {
-        private IIdentityServerConfiguration _identityServerConfig;
-
         public Startup(IHostingEnvironment env)
         {
             var builder = new ConfigurationBuilder()
@@ -46,29 +39,7 @@ namespace Ocelot.ManualTest
             };
 
             services.AddOcelotOutputCaching(settings);
-
-            var username = Environment.GetEnvironmentVariable("OCELOT_USERNAME");
-            var hash = Environment.GetEnvironmentVariable("OCELOT_HASH");
-            var salt = Environment.GetEnvironmentVariable("OCELOT_SALT");
-
-            _identityServerConfig = new IdentityServerConfiguration(
-                "admin",
-                false,
-                SupportedTokens.Both,
-                "secret",
-                new List<string> {"admin", "openid", "offline_access"},
-                "Ocelot Administration",
-                true,
-                GrantTypes.ResourceOwnerPassword,
-                AccessTokenType.Jwt,
-                false,
-                new List<User> 
-                {
-                    new User("admin", username, hash, salt)
-                }
-            );
-
-            services.AddOcelot(Configuration, _identityServerConfig);
+            services.AddOcelot(Configuration);
         }
 
         public async void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
